address nits. split out sui-pg-db into a lib.rs that defines t he concrete Connection and Db, a model.rs for StoredWatermark, and a pg_store.rs that implements the stor e traits
