+++ crates/sui-indexer-alt/src/handlers/cp_sequence_numbers.rs
@@ -8,7 +8,7 @@ use anyhow::{bail, Result};
 use diesel::prelude::*;
 use diesel_async::RunQueryDsl;
 use sui_indexer_alt_framework::{
-    db::{Db, DbConnection},
+    db::{Connection, Db},
     pipeline::{concurrent::Handler, Processor},
     types::full_checkpoint_content::CheckpointData,
 };
@@ -40,7 +40,7 @@ impl Processor for CpSequenceNumbers {
 impl Handler for CpSequenceNumbers {
     type Store = Db;
 
-    async fn commit<'a>(values: &[Self::Value], conn: &mut DbConnection<'a>) -> Result<usize> {
+    async fn commit<'a>(values: &[Self::Value], conn: &mut Connection<'a>) -> Result<usize> {
         Ok(diesel::insert_into(cp_sequence_numbers::table)
             .values(values)
             .on_conflict_do_nothing()
@@ -50,7 +50,7 @@ impl Handler for CpSequenceNumbers {
 }
 
 /// Inclusive start and exclusive end range of prunable txs.
-pub async fn tx_interval(conn: &mut DbConnection<'_>, cps: Range<u64>) -> Result<Range<u64>> {
+pub async fn tx_interval(conn: &mut Connection<'_>, cps: Range<u64>) -> Result<Range<u64>> {
     let result = get_range(conn, cps).await?;
 
     Ok(Range {
@@ -61,7 +61,7 @@ pub async fn tx_interval(conn: &mut DbConnection<'_>, cps: Range<u64>) -> Result
 
 /// Returns the epochs of the given checkpoint range. `start` is the epoch of the first checkpoint
 /// and `end` is the epoch of the last checkpoint.
-pub async fn epoch_interval(conn: &mut DbConnection<'_>, cps: Range<u64>) -> Result<Range<u64>> {
+pub async fn epoch_interval(conn: &mut Connection<'_>, cps: Range<u64>) -> Result<Range<u64>> {
     let result = get_range(conn, cps).await?;
 
     Ok(Range {
@@ -75,7 +75,7 @@ pub async fn epoch_interval(conn: &mut DbConnection<'_>, cps: Range<u64>) -> Res
 /// The values are expected to exist since the cp_sequence_numbers table must have enough information to
 /// encompass the retention of other tables.
 pub(crate) async fn get_range(
-    conn: &mut DbConnection<'_>,
+    conn: &mut Connection<'_>,
     cps: Range<u64>,
 ) -> Result<(StoredCpSequenceNumbers, StoredCpSequenceNumbers)> {
     let Range {
