+++ crates/sui-indexer-alt/src/handlers/ev_emit_mod.rs
@@ -8,7 +8,7 @@ use anyhow::Result;
 use diesel::{ExpressionMethods, QueryDsl};
 use diesel_async::RunQueryDsl;
 use sui_indexer_alt_framework::{
-    db::{Db, DbConnection},
+    db::{Connection, Db},
     pipeline::{concurrent::Handler, Processor},
     types::full_checkpoint_content::CheckpointData,
 };
@@ -58,7 +58,7 @@ impl Handler for EvEmitMod {
     const MIN_EAGER_ROWS: usize = 100;
     const MAX_PENDING_ROWS: usize = 10000;
 
-    async fn commit<'a>(values: &[Self::Value], conn: &mut DbConnection<'a>) -> Result<usize> {
+    async fn commit<'a>(values: &[Self::Value], conn: &mut Connection<'a>) -> Result<usize> {
         Ok(diesel::insert_into(ev_emit_mod::table)
             .values(values)
             .on_conflict_do_nothing()
@@ -70,7 +70,7 @@ impl Handler for EvEmitMod {
         &self,
         from: u64,
         to_exclusive: u64,
-        conn: &mut DbConnection<'a>,
+        conn: &mut Connection<'a>,
     ) -> Result<usize> {
         let Range {
             start: from_tx,
@@ -99,7 +99,7 @@ mod tests {
 
     // A helper function to return all entries in the ev_emit_mod table sorted by package, module,
     // tx_sequence_number, and sender.
-    async fn get_all_ev_emit_mod(conn: &mut DbConnection<'_>) -> Result<Vec<StoredEvEmitMod>> {
+    async fn get_all_ev_emit_mod(conn: &mut Connection<'_>) -> Result<Vec<StoredEvEmitMod>> {
         let query = ev_emit_mod::table
             .order_by((
                 ev_emit_mod::tx_sequence_number,
