+++ crates/sui-indexer-alt/src/handlers/kv_checkpoints.rs
@@ -7,7 +7,7 @@ use anyhow::{Context, Result};
 use diesel::{ExpressionMethods, QueryDsl};
 use diesel_async::RunQueryDsl;
 use sui_indexer_alt_framework::{
-    db::{Db, DbConnection},
+    db::{Connection, Db},
     pipeline::{concurrent::Handler, Processor},
     types::full_checkpoint_content::CheckpointData,
 };
@@ -40,7 +40,7 @@ impl Processor for KvCheckpoints {
 impl Handler for KvCheckpoints {
     type Store = Db;
 
-    async fn commit<'a>(values: &[Self::Value], conn: &mut DbConnection<'a>) -> Result<usize> {
+    async fn commit<'a>(values: &[Self::Value], conn: &mut Connection<'a>) -> Result<usize> {
         Ok(diesel::insert_into(kv_checkpoints::table)
             .values(values)
             .on_conflict_do_nothing()
@@ -52,7 +52,7 @@ impl Handler for KvCheckpoints {
         &self,
         from: u64,
         to_exclusive: u64,
-        conn: &mut DbConnection<'a>,
+        conn: &mut Connection<'a>,
     ) -> Result<usize> {
         let filter = kv_checkpoints::table
             .filter(kv_checkpoints::sequence_number.between(from as i64, to_exclusive as i64 - 1));
@@ -70,7 +70,7 @@ mod tests {
     };
     use sui_indexer_alt_schema::MIGRATIONS;
 
-    async fn get_all_kv_checkpoints(conn: &mut DbConnection<'_>) -> Result<Vec<StoredCheckpoint>> {
+    async fn get_all_kv_checkpoints(conn: &mut Connection<'_>) -> Result<Vec<StoredCheckpoint>> {
         let query = kv_checkpoints::table.load(conn).await?;
         Ok(query)
     }
