+++ crates/sui-indexer-alt/src/handlers/kv_epoch_starts.rs
@@ -8,7 +8,7 @@ use anyhow::{bail, Context, Result};
 use diesel::{ExpressionMethods, QueryDsl};
 use diesel_async::RunQueryDsl;
 use sui_indexer_alt_framework::{
-    db::{Db, DbConnection},
+    db::{Connection, Db},
     pipeline::{concurrent::Handler, Processor},
     types::{
         full_checkpoint_content::CheckpointData,
@@ -73,7 +73,7 @@ impl Handler for KvEpochStarts {
 
     const MIN_EAGER_ROWS: usize = 1;
 
-    async fn commit<'a>(values: &[Self::Value], conn: &mut DbConnection<'a>) -> Result<usize> {
+    async fn commit<'a>(values: &[Self::Value], conn: &mut Connection<'a>) -> Result<usize> {
         Ok(diesel::insert_into(kv_epoch_starts::table)
             .values(values)
             .on_conflict_do_nothing()
@@ -85,7 +85,7 @@ impl Handler for KvEpochStarts {
         &self,
         from: u64,
         to_exclusive: u64,
-        conn: &mut DbConnection<'a>,
+        conn: &mut Connection<'a>,
     ) -> Result<usize> {
         let Range {
             start: from_epoch,
