+++ crates/sui-indexer-alt/src/handlers/obj_versions.rs
@@ -6,7 +6,7 @@ use std::sync::Arc;
 use anyhow::Result;
 use diesel_async::RunQueryDsl;
 use sui_indexer_alt_framework::{
-    db::{Db, DbConnection},
+    db::{Connection, Db},
     pipeline::{concurrent::Handler, Processor},
     types::{effects::TransactionEffectsAPI, full_checkpoint_content::CheckpointData},
 };
@@ -55,7 +55,7 @@ impl Handler for ObjVersions {
     const MIN_EAGER_ROWS: usize = 100;
     const MAX_PENDING_ROWS: usize = 10000;
 
-    async fn commit<'a>(values: &[Self::Value], conn: &mut DbConnection<'a>) -> Result<usize> {
+    async fn commit<'a>(values: &[Self::Value], conn: &mut Connection<'a>) -> Result<usize> {
         Ok(diesel::insert_into(obj_versions::table)
             .values(values)
             .on_conflict_do_nothing()
